+++ session/raw.go
@@ -2,18 +2,29 @@ package session
 
 import (
 	"database/sql"
+	"example.com/mark/geeorm/dialect"
 	"example.com/mark/geeorm/log"
+	"example.com/mark/geeorm/schema"
 	"strings"
 )
 
 type Session struct {
-	db      *sql.DB
-	sql     strings.Builder
-	sqlVars []interface{}
+	db       *sql.DB
+	dialect  dialect.Dialect
+	refTable *schema.Schema
+	sql      strings.Builder
+	sqlVars  []interface{}
 }
 
-func New(db *sql.DB) *Session {
-	return &Session{db: db}
+func New(db *sql.DB, dialect dialect.Dialect) *Session {
+	return &Session{db: db, dialect: dialect}
+}
+
+func (s *Session) Raw(sql string, values ...interface{}) *Session {
+	s.sql.WriteString(sql)
+	s.sql.WriteString(" ")
+	s.sqlVars = append(s.sqlVars, values...)
+	return s
 }
 
 func (s *Session) Clear() {
@@ -23,13 +34,6 @@ func (s *Session) Clear() {
 
 func (s *Session) DB() *sql.DB {
 	return s.db
-}
-
-func (s *Session) Raw(sql string, values ...interface{}) *Session {
-	s.sql.WriteString(sql)
-	s.sql.WriteString(" ")
-	s.sqlVars = append(s.sqlVars, values...)
-	return s
 }
 
 func (s *Session) Exec() (result sql.Result, err error) {