+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 1.0.0 (2021-06-03)
+
+
+### Features
+
+* **dialect:** add support to multiDB dialect ([5335347](https://github.com/runmark/orm-7-days/commit/53353478608741954e446e1f82e91e5255fe0033))
+* **geeorm:** 7 day a geeorm  initialization ([48adb2e](https://github.com/runmark/orm-7-days/commit/48adb2e2131557f9576a14046470c54ece2bfc4f))
+* **schema:** use schema to convert between go struct and db table ([7f98723](https://github.com/runmark/orm-7-days/commit/7f987232157ecc39a84d039238c829e0a82c6abb))
+
+
+
 # 1.0.0 (2021-06-02)